env: do not scan package if its not exist on local
Previously, when freezing the package and package did not exist on local
file system, we do scan the package for version and remote URL, which
cause an error because the package directory did not exist.

This commit fix the issue by not scanning the package if its not exist on
local system.