all: replace beku_test in testdata with gitsubmodule
Previously, the beku_test is bare clone of git repository stored as is.
This cause an initial clone of the beku repository itself does not
recognize it as git repository, which make the test fail:

  go test -v -coverprofile=cover.out ./... || rm -f cover.out
  fatal: '/home/ms/go/src/git.sr.ht/~shulhan/beku/testdata/beku_test.git' does not appear to be a git repository
  fatal: Could not read from remote repository.

  Please make sure you have the correct access rights
  and the repository exists.
  2022/01/30 10:56:19 gitInstall: Clone: exit status 128
  FAIL    github.com/shuLhan/beku 0.008s

This changes make the testdata/beku_test.git as git submodule, to make
the test run well.